requirements.txt
aiodns==3.1.1
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio @ file:///home/conda/feedstock_root/build_artifacts/anyio_1700835416766/work
attrs==23.1.0
beautifulsoup4==4.12.2
blessed==1.20.0
blinker==1.7.0
blis==0.7.11
Brotli==1.1.0
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cryptography==41.0.7
cymem==2.0.8
de-core-news-md @ https://github.com/explosion/spacy-models/releases/download/de_core_news_md-3.7.0/de_core_news_md-3.7.0-py3-none-any.whl#sha256=1426896f135b7e6314637faa9025a3e580c9ba2785372ccffe723dc20b41c493
deep-translator==1.11.4
defusedxml==0.7.1
dependency-injector==4.41.0
duckduckgo-search==3.9.10
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup @ file:///work/ci_py311/exceptiongroup_1676827727458/work
extruct==0.16.0
fastapi==0.105.0
Flask==3.0.0
frozenlist==1.4.1
future==0.18.3
getuseragent==0.0.7
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html-text==0.5.2
html5lib==1.1
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
humanize==4.9.0
hyperframe==6.0.1
idna @ file:///work/ci_py311/idna_1676822698822/work
ingredient-parser-nlp==0.1.0b7
inquirer==3.1.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jstyleson==0.0.2
kaitaistruct==0.10
langcodes==3.3.0
lidl-plus==0.3.4
lxml==4.9.3
Mako==1.3.0
MarkupSafe==2.1.3
mf2py==2.0.1
multidict==6.0.4
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.2
oic==1.6.1
openfoodfacts==0.1.12
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
patsy==0.5.5
Pint==0.22
preshed==3.0.9
pyasn1==0.5.1
pycares==4.4.0
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyjwkest==1.4.2
pyOpenSSL==23.3.0
pyparsing==3.1.1
pyRdfa3==3.5.3
PySocks==1.7.1
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
pytz==2023.3.post1
PyYAML==6.0.1
rdflib==7.0.0
readchar==4.0.5
regex==2023.10.3
requests==2.31.0
schedule==1.2.1
scipy==1.11.4
scrape-schema-recipe==0.2.2
selenium==4.16.0
selenium-wire==5.1.0
six==1.16.0
smart-open==6.4.0
sniffio @ file:///work/ci_py311/sniffio_1676823742530/work
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.27.0
statsmodels==0.14.1
thinc==8.2.1
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
w3lib==2.1.2
wasabi==1.1.2
watchfiles==0.21.0
weasel==0.3.4
webdriver-manager==4.0.1
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.1
Whoosh==2.7.4
wsproto==1.2.0
yarl==1.9.4
zstandard==0.22.0
